chat/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==7.0.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==39.0.1
daphne==3.0.2
Django==3.2.17
django-channels-graphql-ws==0.9.1
django-cors-headers==3.13.0
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
frozenlist==1.3.3
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.4
multidict==6.0.4
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.0.0
pytz==2022.7.1
Rx==1.6.3
service-identity==21.1.0
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
yarl==1.8.2
zope.interface==5.5.2
